+++ eoreader/products/optical/s3_slstr_product.py
@@ -27,11 +27,12 @@ from enum import unique
 from functools import reduce
 from typing import Union
 
+import geopandas as gpd
 import numpy as np
 import xarray as xr
 from rasterio import features
 from rasterio.enums import Resampling
-from sertit import path, rasters, types
+from sertit import files, path, rasters, types
 from sertit.misc import ListEnum
 from sertit.types import AnyPathStrType, AnyPathType
 
@@ -58,7 +59,13 @@ from eoreader.bands import (
     to_str,
 )
 from eoreader.exceptions import InvalidTypeError
-from eoreader.keywords import CLEAN_OPTICAL, SLSTR_RAD_ADJUST, SLSTR_STRIPE, SLSTR_VIEW
+from eoreader.keywords import (
+    CLEAN_OPTICAL,
+    SLSTR_RAD_ADJUST,
+    SLSTR_STRIPE,
+    SLSTR_VIEW,
+    TO_REFLECTANCE,
+)
 from eoreader.products import S3DataType, S3Product, S3ProductType
 from eoreader.products.optical.optical_product import DEF_CLEAN_METHOD, CleanMethod
 from eoreader.stac import ASSET_ROLE, BT, CENTER_WV, DESCRIPTION, FWHM, GSD, ID, NAME
@@ -1162,6 +1169,24 @@ class S3SlstrProduct(S3Product):
         suffix = self._get_suffix(band, **kwargs)
         res_str = self._pixel_size_to_str(pixel_size)
 
+        # Window name
+        window = kwargs.get("window")
+
+        win_suffix = ""
+        if window is not None:
+            if path.is_path(window):
+                win_suffix = f"{path.get_filename(window)}"
+            elif isinstance(window, gpd.GeoDataFrame):
+                win_suffix = f"{window.attrs.get('name')}"
+
+            if not win_suffix:
+                win_suffix = f"win{files.hash_file_content(str(window))}"
+
+            win_suffix += "_"
+
+        # Radiometric processing
+        rad_proc = "" if kwargs.get(TO_REFLECTANCE, True) else "_as_is"
+
         return self._get_band_folder(writable).joinpath(
-            f"{self.condensed_name}_{band.name}_{suffix}_{res_str.replace('.', '-')}_{cleaning_method.value}.tif",
+            f"{self.condensed_name}_{band.name}_{suffix}_{res_str.replace('.', '-')}_{win_suffix}{cleaning_method.value}{rad_proc}.tif",
         )